requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.19.1
gradio_client==0.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.14
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==3.11.0
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycairo==1.23.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.44.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
semantic-version==2.10.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.7.0
snowflake-snowpark-python==1.12.1
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
stack-data==0.6.2
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tomlkit==0.12.0
toolz==0.12.1
tornado==6.1
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.27.1
wcwidth==0.2.13
websockets==11.0.3
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
